Move PlantDatabase next to IDatabase. Fix singleton constructor
