+++ Plants/APlant.cs
@@ -2,7 +2,7 @@ using System;
 using System.Collections.Generic;
 using Godot;
 using untitledplantgame.Common;
-using untitledplantgame.ResourceData.Resources.Plants;
+using untitledplantgame.ResourceData;
 
 namespace untitledplantgame.Plants;
