+++ ResourceData/IDatabase.cs
@@ -1,5 +1,4 @@
-﻿using System.Collections.Generic;
-using Godot;
+﻿using Godot;
 
 namespace untitledplantgame.ResourceData;
 
@@ -9,12 +8,23 @@ public interface IDatabase<T> where T: Resource
 	// e.g. "res://Resources/Items" for ItemDatabase or "res://Resources/Dialoue" for DialogueDatabase
 	string DirPath { get; set; } //TODO do we need this?
 	
-	//Get resource of type T by name
+	/// <summary>
+	/// Get resource of type T by name. If no resource is found, return null.
+	/// </summary>
+	/// <param name="name"></param>
+	/// <returns></returns>
 	T GetResourceByName(string name);
 	
-	//Get resource of type T by id
+	/// <summary>
+	/// Get resource of type T by id. If no resource is found, return null.
+	/// </summary>
+	/// <param name="id"></param>
+	/// <returns></returns>
 	T GetResourceById(int id);
 	
-	//Get all resources of type T
+	/// <summary>
+	/// Get all resources of type T. If no resources are found, return an empty array.
+	/// </summary>
+	/// <returns></returns>
 	T[] GetAllResources();
 }
