+++ ResourceData/PlantDatabase.cs
@@ -1,9 +1,14 @@
 using System.Linq;
 using Godot;
+using untitledplantgame.Common;
 using untitledplantgame.Plants;
 
-namespace untitledplantgame.ResourceData.Resources.Plants;
+namespace untitledplantgame.ResourceData;
 
+
+/// <summary>
+/// A database singleton for all plant resources.
+/// </summary>
 public partial class PlantDatabase : Node, IDatabase<PlantData>
 {
 	public static PlantDatabase Instance { get; private set; }
@@ -13,11 +18,21 @@ public partial class PlantDatabase : Node, IDatabase<PlantData>
 		get => _dirPath;
 		set => _dirPath = value;
 	}
-	
+
+	private Logger _logger;
 	private string _dirPath = "res://ResourceData/Resources/Plants";
 
 	public override void _Ready()
 	{
+		_logger = new Logger(this);
+		
+		if (Instance != null)
+		{
+			_logger.Error("There are multiple PlantDatabases in the scene. There should only be one.");
+			QueueFree();
+			return;
+		}
+		
 		Instance = this;
 	}
