+++ project.godot
@@ -29,9 +29,9 @@ config/icon="res://icon.svg"
 
 GameStateMachine="*res://Common/GameState/GameStateMachine.cs"
 EventBus="*res://EventBus/EventBus.cs"
-ResourceManager="*res://ResourceData/ResourceManager.cs"
 TimeController="*res://Common/TimeController.cs"
 WeatherCycle="*res://Cycle/Weather/WeatherCycle.cs"
+PlantDatabase="*res://ResourceData/PlantDatabase.cs"
 
 [display]
